Fix selections + commands when wysiwyg editing with special characters
Converting back and forth to a native `Selection` object required accommodating the HTML special characters in the offsets. The string the browser returns from `textContent` is not escaped which is a problem when trying to work with a selection in a block with a special character around.

This should finish out the fix for bevacqua/woofmark#47