Dependencies Trimmed Down & Moved to Bower
Gonna make this a proper bower module hopefully. Things are going to start to look very different

TODO: `beauracracy` is still not on bower, but I think I'm going to remove that code for now. There is also the possibilty of pull pre-4.2 code from woofmark, which didn't have that dependency