Marking a Full Split from woofmark. Rename to barkmark
We're going to start to do some significant changes to the code so I want to split things off so there is no confusion. I'll be reworking or rewriting some components as we need and also likely removing some pieces and functionality. Don't want any confusion over the differences.

Note that we're keeping in place all license and will continue to point towards woofmark quite a bit likely for a while. If there comes such a time as we do a *full* rewrite we will split off into a new repository and break from the forked repo here. But until such time we will keep the ties to the original and, if possible, try to send some of our more applicable changes upstream.