Split up Surface & Editor classes
Each mode now has it's own surface with a standard interface. Will expand upon that going forward for extensible modes but still a ways off from that.

Splitting out the editor from barkmark root file allows there to also be a manager and for us to treat that base object as a wrapper where other things can occur. It also allows us some freedom in the future for an API that controls the managers and editors.